Rename file to match the func names
